reorganize multiple classes into different files
